Move helper/spec_helper.rb to test/
