+++ test/http.rb
@@ -1,4 +1,4 @@
-require_relative "../helper/spec_helper.rb"
+require_relative "spec_helper"
 
 describe "SLES 12 HTTP server " do
   it "checks, apache2 and firewall is running correctly" do
