+++ test/sles12_mini_network.rb
@@ -1,4 +1,4 @@
-require_relative "../helper/spec_helper.rb"
+require_relative "spec_helper"
 
 describe "SLES 12 first stage only," do
   it "if user -vagrant- has been created" do
