+++ test/tftp.rb
@@ -1,4 +1,4 @@
-require_relative "../helper/spec_helper.rb"
+require_relative "spec_helper"
 
 describe "SLES 12 TFTP server " do
   it "checks, tftp is running correctly" do
