+++ test/upgrade_sles11.rb
@@ -1,4 +1,4 @@
-require_relative "../helper/spec_helper.rb"
+require_relative "spec_helper"
 
 describe "SLES11 - SLES12 upgrade checks," do
   it "if user -vagrant- has been created" do
